Overrode the colour class' eq function so that it recognises instances of the colour class from the engine module as identical.
Engine: Improved the evaluation function to include piece-square evaluation and sign differentiation. Small changes to the Zobrist generating and hashing functions. Created an alpha-beta searching function, and a NodeType enum to distinguish between the different nodes. Started to write a function which would find the optimal move using the alpha-beta search.